requirements.txt
reportlab==4.2.5
PyPDF2==3.0.1
Pillow==11.1.0
django-crispy-forms
django
crispy-bootstrap5
python-decouple
num2words